+++ photon-api/src/main/scala/com/linkedin/photon/ml/SparkSessionConfiguration.scala
@@ -29,7 +29,7 @@ import com.linkedin.photon.ml.normalization.NormalizationContext
 import com.linkedin.photon.ml.optimization._
 import com.linkedin.photon.ml.optimization.game.GLMOptimizationConfiguration
 import com.linkedin.photon.ml.projector.LinearSubspaceProjector
-import com.linkedin.photon.ml.supervised.classification.{LogisticRegressionModel, SmoothedHingeLossLinearSVMModel}
+import com.linkedin.photon.ml.supervised.classification.LogisticRegressionModel
 import com.linkedin.photon.ml.supervised.model.GeneralizedLinearModel
 import com.linkedin.photon.ml.supervised.regression.{LinearRegressionModel, PoissonRegressionModel}
 
@@ -70,7 +70,6 @@ object SparkSessionConfiguration {
     classOf[Set[Int]],
     classOf[SingleNodeObjectiveFunction],
     classOf[SingleNodeOptimizationProblem[_]],
-    classOf[SmoothedHingeLossLinearSVMModel],
     classOf[SparseVector[Double]],
     classOf[TRON],
     classOf[ValueAndGradientAggregator],
