+++ photon-api/src/main/scala/com/linkedin/photon/ml/evaluation/EvaluatorFactory.scala
@@ -47,8 +47,6 @@ object EvaluatorFactory {
 
       case LogisticLoss => LogisticLossEvaluator
 
-      case SmoothedHingeLoss => SmoothedHingeLossEvaluator
-
       case SquaredLoss => SquaredLossEvaluator
 
       case MultiPrecisionAtK(k, idTag) =>
