+++ photon-api/src/test/scala/com/linkedin/photon/ml/function/ObjectiveFunctionHelperTest.scala
@@ -14,16 +14,13 @@
  */
 package com.linkedin.photon.ml.function
 
+import org.mockito.Mockito._
 import org.testng.Assert._
 import org.testng.annotations.{DataProvider, Test}
 
 import com.linkedin.photon.ml.TaskType
-import com.linkedin.photon.ml.TaskType.TaskType
-import com.linkedin.photon.ml.function.glm.DistributedGLMLossFunction
-import com.linkedin.photon.ml.function.svm.DistributedSmoothedHingeLossFunction
+import com.linkedin.photon.ml.optimization.game.{CoordinateOptimizationConfiguration, RandomEffectOptimizationConfiguration}
 import com.linkedin.photon.ml.optimization.game.FixedEffectOptimizationConfiguration
-import com.linkedin.photon.ml.optimization.{OptimizerConfig, OptimizerType}
-import com.linkedin.photon.ml.supervised.model.GeneralizedLinearModel
 
 /**
  * Unit tests for [[ObjectiveFunctionHelper]].
@@ -35,42 +32,34 @@ class ObjectiveFunctionHelperTest {
   @DataProvider
   def trainingTaskProvider(): Array[Array[Any]] =
     Array(
-      Array(TaskType.LOGISTIC_REGRESSION),
-      Array(TaskType.LINEAR_REGRESSION),
-      Array(TaskType.POISSON_REGRESSION),
-      Array(TaskType.SMOOTHED_HINGE_LOSS_LINEAR_SVM))
+      Array(FIXED_COORDINATE_OPT_CONFIG),
+      Array(RANDOM_COORDINATE_OPT_CONFIG))
 
   /**
    * Test that the [[ObjectiveFunction]] generated by the factory function returned by the [[ObjectiveFunctionHelper]]
-   * is of the appropriate type for the given training task.
+   * is of the appropriate type for the given coordinate configuration.
    *
-   * @param trainingTask The training task
+   * @param config The coordinate configuration
    */
   @Test(dataProvider = "trainingTaskProvider")
-  def testBuildFactory(trainingTask: TaskType): Unit = {
+  def testBuildFactory(config: CoordinateOptimizationConfiguration): Unit = {
 
-    val objectiveFunction = ObjectiveFunctionHelper.buildFactory(
-      trainingTask,
-      TREE_AGGREGATE_DEPTH)(COORDINATE_OPT_CONFIG)
+    val objectiveFunctionFactory =
+      ObjectiveFunctionHelper.buildFactory(TaskType.LOGISTIC_REGRESSION, TREE_AGGREGATE_DEPTH)(config)
 
-    trainingTask match {
-      case TaskType.LOGISTIC_REGRESSION | TaskType.LINEAR_REGRESSION | TaskType.POISSON_REGRESSION =>
-        assertTrue(
-          objectiveFunction.isInstanceOf[(Option[GeneralizedLinearModel], Option[Int]) => DistributedGLMLossFunction])
+    config match {
+      case _: FixedEffectOptimizationConfiguration =>
+        assertTrue(objectiveFunctionFactory.isInstanceOf[ObjectiveFunctionHelper.DistributedObjectiveFunctionFactory])
 
-      case TaskType.SMOOTHED_HINGE_LOSS_LINEAR_SVM =>
-        assertTrue(
-          objectiveFunction
-            .isInstanceOf[(Option[GeneralizedLinearModel], Option[Int]) => DistributedSmoothedHingeLossFunction])
+      case _: RandomEffectOptimizationConfiguration =>
+        assertTrue(objectiveFunctionFactory.isInstanceOf[ObjectiveFunctionHelper.SingleNodeObjectiveFunctionFactory])
     }
   }
 }
 
 object ObjectiveFunctionHelperTest {
 
-  val COORDINATE_OPT_CONFIG = FixedEffectOptimizationConfiguration(OptimizerConfig(OptimizerType.LBFGS, 1, 2e-2))
-  val ENABLE_INCREMENTAL_TRAINING = false
-  val MAXIMUM_ITERATIONS = 1
-  val TOLERANCE = 2e-2
-  val TREE_AGGREGATE_DEPTH = 3
+  private val FIXED_COORDINATE_OPT_CONFIG = mock(classOf[FixedEffectOptimizationConfiguration])
+  private val RANDOM_COORDINATE_OPT_CONFIG = mock(classOf[RandomEffectOptimizationConfiguration])
+  private val TREE_AGGREGATE_DEPTH = 3
 }
