+++ photon-api/src/test/scala/com/linkedin/photon/ml/optimization/GeneralizedLinearOptimizationProblemTest.scala
@@ -14,21 +14,17 @@
  */
 package com.linkedin.photon.ml.optimization
 
-import scala.math.abs
-
-import breeze.linalg.{Vector, sum}
+import breeze.linalg.Vector
 import org.mockito.Mockito._
 import org.testng.Assert._
 import org.testng.annotations.Test
 
 import com.linkedin.photon.ml.data.LabeledPoint
 import com.linkedin.photon.ml.function._
-import com.linkedin.photon.ml.function.svm.SingleNodeSmoothedHingeLossFunction
 import com.linkedin.photon.ml.model.Coefficients
-import com.linkedin.photon.ml.supervised.classification.{LogisticRegressionModel, SmoothedHingeLossLinearSVMModel}
+import com.linkedin.photon.ml.supervised.classification.LogisticRegressionModel
 import com.linkedin.photon.ml.supervised.model.GeneralizedLinearModel
 import com.linkedin.photon.ml.supervised.regression.{LinearRegressionModel, PoissonRegressionModel}
-import com.linkedin.photon.ml.test.CommonTestUtils
 import com.linkedin.photon.ml.test.CommonTestUtils.generateDenseVector
 
 /**
@@ -42,7 +38,7 @@ class GeneralizedLinearOptimizationProblemTest {
   def testInitializeZeroModel(): Unit = {
     val optimizer = mock(classOf[Optimizer[ObjectiveFunction]])
     val statesTracker = mock(classOf[OptimizationStatesTracker])
-    val objective = mock(classOf[SingleNodeSmoothedHingeLossFunction])
+    val objective = mock(classOf[SingleNodeObjectiveFunction])
     val regularization = NoRegularizationContext
 
     doReturn(statesTracker).when(optimizer).getStateTracker
@@ -62,16 +58,10 @@ class GeneralizedLinearOptimizationProblemTest {
       objective,
       PoissonRegressionModel.apply,
       regularization)
-    val hingeProblem = new MockOptimizationProblem(
-      optimizer,
-      objective,
-      SmoothedHingeLossLinearSVMModel.apply,
-      regularization)
 
     val logisticModel = logisticProblem.publicInitializeZeroModel(DIMENSION)
     val linearModel = linearProblem.publicInitializeZeroModel(DIMENSION)
     val poissonModel = poissonProblem.publicInitializeZeroModel(DIMENSION)
-    val hingeModel = hingeProblem.publicInitializeZeroModel(DIMENSION)
 
     assertTrue(logisticModel.isInstanceOf[LogisticRegressionModel])
     assertEquals(logisticModel.coefficients.means.length, DIMENSION)
@@ -81,16 +71,14 @@ class GeneralizedLinearOptimizationProblemTest {
 
     assertTrue(poissonModel.isInstanceOf[PoissonRegressionModel])
     assertEquals(poissonModel.coefficients.means.length, DIMENSION)
-
-    assertTrue(hingeModel.isInstanceOf[SmoothedHingeLossLinearSVMModel])
-    assertEquals(hingeModel.coefficients.means.length, DIMENSION)
   }
 
   @Test
   def testCreateModel(): Unit = {
+
     val optimizer = mock(classOf[Optimizer[ObjectiveFunction]])
     val statesTracker = mock(classOf[OptimizationStatesTracker])
-    val objective = mock(classOf[SingleNodeSmoothedHingeLossFunction])
+    val objective = mock(classOf[SingleNodeObjectiveFunction])
     val regularization = NoRegularizationContext
 
     doReturn(statesTracker).when(optimizer).getStateTracker
@@ -110,17 +98,11 @@ class GeneralizedLinearOptimizationProblemTest {
       objective,
       PoissonRegressionModel.apply,
       regularization)
-    val hingeProblem = new MockOptimizationProblem(
-      optimizer,
-      objective,
-      SmoothedHingeLossLinearSVMModel.apply,
-      regularization)
     val coefficients = generateDenseVector(DIMENSION)
 
     val logisticModel = logisticProblem.publicCreateModel(coefficients, None)
     val linearModel = linearProblem.publicCreateModel(coefficients, None)
     val poissonModel = poissonProblem.publicCreateModel(coefficients, None)
-    val hingeModel = hingeProblem.publicCreateModel(coefficients, None)
 
     assertTrue(logisticModel.isInstanceOf[LogisticRegressionModel])
     assertEquals(coefficients, logisticModel.coefficients.means)
@@ -130,70 +112,6 @@ class GeneralizedLinearOptimizationProblemTest {
 
     assertTrue(poissonModel.isInstanceOf[PoissonRegressionModel])
     assertEquals(coefficients, poissonModel.coefficients.means)
-
-    assertTrue(hingeModel.isInstanceOf[SmoothedHingeLossLinearSVMModel])
-    assertEquals(coefficients, hingeModel.coefficients.means)
-  }
-
-  @Test
-  def testGetRegularizationTermValue(): Unit = {
-    val coefficients = new Coefficients(generateDenseVector(DIMENSION))
-    val regWeight = 10D
-    val alpha = 0.25
-    val l1RegWeight = alpha * regWeight
-    val l2RegWeight = (1 - alpha) * regWeight
-    val expectedL1Term = sum(coefficients.means.map(abs)) * l1RegWeight
-    val expectedL2Term = coefficients.means.dot(coefficients.means) * l2RegWeight / 2.0
-    val expectedElasticNetTerm = expectedL1Term + expectedL2Term
-
-    val optimizerNoReg = mock(classOf[LBFGS])
-    val optimizerL1Reg = mock(classOf[OWLQN])
-    val objectiveNoReg = mock(classOf[SingleNodeSmoothedHingeLossFunction])
-    val objectiveL2Reg = mock(classOf[L2LossFunction])
-    val statesTracker = mock(classOf[OptimizationStatesTracker])
-    val initialModel = mock(classOf[GeneralizedLinearModel])
-
-    doReturn(statesTracker).when(optimizerNoReg).getStateTracker
-    doReturn(statesTracker).when(optimizerL1Reg).getStateTracker
-
-    val problemNone = new MockOptimizationProblem(
-      optimizerNoReg,
-      objectiveNoReg,
-      LogisticRegressionModel.apply,
-      NoRegularizationContext)
-    val problemL1 = new MockOptimizationProblem(
-      optimizerL1Reg,
-      objectiveNoReg,
-      LogisticRegressionModel.apply,
-      L1RegularizationContext)
-    val problemL2 = new MockOptimizationProblem(
-      optimizerNoReg,
-      objectiveL2Reg,
-      LogisticRegressionModel.apply,
-      L2RegularizationContext)
-    val problemElasticNet = new MockOptimizationProblem(
-      optimizerL1Reg,
-      objectiveL2Reg,
-      LogisticRegressionModel.apply,
-      ElasticNetRegularizationContext(alpha))
-
-    doReturn(l1RegWeight).when(optimizerL1Reg).l1RegularizationWeight
-    doReturn(l2RegWeight).when(objectiveL2Reg).l2RegularizationWeight
-    doReturn(coefficients).when(initialModel).coefficients
-
-    assertEquals(0.0, problemNone.getRegularizationTermValue(initialModel), CommonTestUtils.HIGH_PRECISION_TOLERANCE)
-    assertEquals(
-      expectedL1Term,
-      problemL1.getRegularizationTermValue(initialModel),
-      CommonTestUtils.HIGH_PRECISION_TOLERANCE)
-    assertEquals(
-      expectedL2Term,
-      problemL2.getRegularizationTermValue(initialModel),
-      CommonTestUtils.HIGH_PRECISION_TOLERANCE)
-    assertEquals(
-      expectedElasticNetTerm,
-      problemElasticNet.getRegularizationTermValue(initialModel),
-      CommonTestUtils.HIGH_PRECISION_TOLERANCE)
   }
 }
 
@@ -202,11 +120,11 @@ object GeneralizedLinearOptimizationProblemTest {
   private val DIMENSION = 10
 
   private class MockOptimizationProblem(
-      optimizer: Optimizer[SingleNodeSmoothedHingeLossFunction],
-      objectiveFunction: SingleNodeSmoothedHingeLossFunction,
+      optimizer: Optimizer[SingleNodeObjectiveFunction],
+      objectiveFunction: SingleNodeObjectiveFunction,
       glmConstructor: Coefficients => GeneralizedLinearModel,
       regularizationContext: RegularizationContext)
-    extends GeneralizedLinearOptimizationProblem[SingleNodeSmoothedHingeLossFunction](
+    extends GeneralizedLinearOptimizationProblem[SingleNodeObjectiveFunction](
       optimizer,
       objectiveFunction,
       glmConstructor,
@@ -251,7 +169,4 @@ object GeneralizedLinearOptimizationProblemTest {
     override def run(input: Iterable[LabeledPoint], initialModel: GeneralizedLinearModel): (GeneralizedLinearModel, OptimizationStatesTracker) =
       (mockGLM, mockStateTracker)
   }
-
-  // No way to pass Mixin class type to Mockito, need to define a concrete class
-  private class L2LossFunction extends SingleNodeSmoothedHingeLossFunction with L2RegularizationDiff
 }
