+++ photon-api/src/test/scala/com/linkedin/photon/ml/sampling/DownSamplerHelperTest.scala
@@ -32,8 +32,7 @@ class DownSamplerHelperTest {
     Array(
       Array(TaskType.LOGISTIC_REGRESSION),
       Array(TaskType.LINEAR_REGRESSION),
-      Array(TaskType.POISSON_REGRESSION),
-      Array(TaskType.SMOOTHED_HINGE_LOSS_LINEAR_SVM))
+      Array(TaskType.POISSON_REGRESSION))
 
   /**
    * Test that the [[DownSampler]] generated by the factory function returned by the [[DownSamplerHelper]] is of the
@@ -52,7 +51,7 @@ class DownSamplerHelperTest {
       case TaskType.LINEAR_REGRESSION | TaskType.POISSON_REGRESSION =>
         assertTrue(downSampler.isInstanceOf[DefaultDownSampler])
 
-      case TaskType.LOGISTIC_REGRESSION | TaskType.SMOOTHED_HINGE_LOSS_LINEAR_SVM =>
+      case TaskType.LOGISTIC_REGRESSION =>
         assertTrue(downSampler.isInstanceOf[BinaryClassificationDownSampler])
     }
   }
