+++ photon-lib/src/main/scala/com/linkedin/photon/ml/TaskType.scala
@@ -21,5 +21,5 @@ object TaskType extends Enumeration {
 
   type TaskType = Value
 
-  val LINEAR_REGRESSION, POISSON_REGRESSION, LOGISTIC_REGRESSION, SMOOTHED_HINGE_LOSS_LINEAR_SVM = Value
+  val LINEAR_REGRESSION, POISSON_REGRESSION, LOGISTIC_REGRESSION = Value
 }
